Fixed errors

+++ image_proc/src/shaders/lowpass.glsl
@@ -3,7 +3,7 @@
 layout (local_size_x = 16, local_size_y = 16, local_size_z = 1) in;
 
 layout (push_constant) uniform PushConstantData {
-        double[9] mask;
+        float[9] mask;
 } dat;
 
 layout (set = 0, binding = 0, rgba8) uniform readonly image2D inImage;
